WIP: the infrastructure is set up to do frame swapping
It's just not drawing certain elements
