+++ qatool/Forget_Password.py
@@ -1,24 +1,22 @@
-import subprocess
 import tkinter as tk
 from tkinter import PhotoImage, messagebox, ttk
 import mysql.connector as mysql
 
-class PasswordResetApp:
-    def __init__(self, root):
-        self.root = root
+import os
+cwd = os.getcwd()
+
+class PasswordResetApp(tk.Frame):
+    def __init__(self, parent, controller):
+        tk.Frame.__init__(self, parent)
+        self.root = controller
         self.root.title('QA Tool : Forgot Password')
-        self.root.geometry('1132x611+300+100')
-        self.root.configure(bg="#fff")
-        self.root.resizable(True, True)
-        img = PhotoImage(file='/Users/viveksonar/Desktop/project_research/iconssss.png')
-        self.root.iconphoto(False, img)
         self.create_password_window = None  # Initialize the reference for the Toplevel window
         self.init_ui()
 
     def show_message_box(self, title, message, delay=1500):
         messagebox.showinfo(title, message)
-        self.root.update()
-        self.root.after(delay, self.root.focus_force)
+        self.update()
+        self.after(delay, self.focus_force)
 
     def get_database_connection(self):
         try:
@@ -57,80 +55,72 @@ class PasswordResetApp:
                 con.close()
 
     def set_new_password(self, username):
-        self.root.withdraw()
-        
+        self.withdraw()
+
         # Create a new Toplevel window
-        self.create_password_window = tk.Toplevel(self.root)
+        self.create_password_window = tk.Toplevel(self)
         self.create_password_window.title("Set New Password")
         self.create_password_window.geometry('1132x611+300+100')
-        
+
         # Create and place the frame within the Toplevel window
         set_password_frame = tk.Frame(self.create_password_window, width=400, height=200, bd=5, relief='solid', borderwidth=1)
         set_password_frame.place(relx=0.5, rely=0.5, anchor=tk.CENTER)
-        
+
         # Create Password label and entry field
         label_createpassword = tk.Label(set_password_frame, text='Create Password', font=('Microsoft YaHei UI Light', 15))
         label_createpassword.grid(row=0, column=0, pady=5, padx=5)
         createpassword_entry = tk.Entry(set_password_frame, width=30, font=('Microsoft YaHei UI Light', 15), bd=1, relief="solid")
         createpassword_entry.grid(row=0, column=1, pady=5, padx=5)
-        
+
         # Confirm Password label and entry field
         label_confirmpassword = tk.Label(set_password_frame, text='Confirm Password', font=('Microsoft YaHei UI Light', 15))
         label_confirmpassword.grid(row=1, column=0, pady=5, padx=5)
         confirmpassword_entry = tk.Entry(set_password_frame, width=30, font=('Microsoft YaHei UI Light', 15), bd=1, relief="solid")
         confirmpassword_entry.grid(row=1, column=1, pady=5, padx=5)
-        
+
         # Function to save the new password
         def save_new_password():
             new_password = createpassword_entry.get()
             confirm_new_password = confirmpassword_entry.get()
-            
+
             if new_password and confirm_new_password and new_password == confirm_new_password:
                 self.update_password_for_username(username, new_password)
                 self.show_message_box("Password Updated", "Your password has been updated successfully.")
                 self.create_password_window.destroy()  # Close the set new password window
-                self.root.destroy()  # Close the forget password window
-                self.open_login_window()  # Close the window upon successful update
+                self.root.show_frame(LoginScreen)
             else:
                 self.show_message_box("Error", "Passwords do not match. Please try again.")
-        
+
         # Set Password button
         set_password_button = ttk.Button(set_password_frame, text="Set Password", command=save_new_password)
         set_password_button.grid(row=2, column=1, columnspan=2, pady=10)
 
     def init_ui(self):
-        self.img = tk.PhotoImage(file='/Users/viveksonar/Desktop/project_research/zoomimagepeopr.png')
-        tk.Label(self.root, image=self.img, bg='white').place(x=0, y=0, relwidth=1, relheight=1)
+        self.img = tk.PhotoImage(file=cwd+'/zoomimagepeopr.png')
+        tk.Label(self, image=self.img, bg='white').place(x=0, y=0, relwidth=1, relheight=1)
 
-        # Frame for the Forgot Password section
-        self.forget_password_frame = tk.Frame(self.root, width=500, height=300, bd=5, relief='solid', borderwidth=1)
-        self.forget_password_frame.place(relx=0.5, rely=0.5, anchor=tk.CENTER)
-
-        self.label_username = tk.Label(self.forget_password_frame, text='Username:', font=('Microsoft YaHei UI Light', 15))
+        self.label_username = tk.Label(self, text='Username:', font=('Microsoft YaHei UI Light', 15))
         self.label_username.grid(row=0, column=0, pady=5, padx=5)
 
-        self.username_entry = tk.Entry(self.forget_password_frame, width=30, font=('Microsoft YaHei UI Light', 15), bd=1, relief="solid")
+        self.username_entry = tk.Entry(self, width=30, font=('Microsoft YaHei UI Light', 15), bd=1, relief="solid")
         self.username_entry.grid(row=0, column=1, pady=5, padx=5)
 
-        self.label_security_question = tk.Label(self.forget_password_frame, text='Select Security Question:', font=('Microsoft YaHei UI Light', 15))
+        self.label_security_question = tk.Label(self, text='Select Security Question:', font=('Microsoft YaHei UI Light', 15))
         self.label_security_question.grid(row=1, column=0, pady=5, padx=5)
 
         self.security_questions = ["What is the name of your first car?", "Which Elementary school did you attend?"]
-        self.security_question_combobox = ttk.Combobox(self.forget_password_frame, values=self.security_questions, width=28, font=('Microsoft YaHei UI Light', 15), state="readonly")
+        self.security_question_combobox = ttk.Combobox(self, values=self.security_questions, width=28, font=('Microsoft YaHei UI Light', 15), state="readonly")
         self.security_question_combobox.grid(row=1, column=1, pady=5, padx=5)
         self.security_question_combobox.current(0)  # Optional: set default value
 
-        self.label_answer = tk.Label(self.forget_password_frame, text='Enter Security Answer:', font=('Microsoft YaHei UI Light', 15))
+        self.label_answer = tk.Label(self, text='Enter Security Answer:', font=('Microsoft YaHei UI Light', 15))
         self.label_answer.grid(row=2, column=0, pady=5, padx=5)
 
-        self.answer_entry = tk.Entry(self.forget_password_frame, width=30, font=('Microsoft YaHei UI Light', 15), bd=1, relief="solid")
+        self.answer_entry = tk.Entry(self, width=30, font=('Microsoft YaHei UI Light', 15), bd=1, relief="solid")
         self.answer_entry.grid(row=2, column=1, pady=5, padx=5)
 
-        self.submit_button = ttk.Button(self.forget_password_frame, text="Submit", command=self.check_answer)
+        self.submit_button = ttk.Button(self, text="Submit", command=self.check_answer)
         self.submit_button.grid(row=3, column=1, pady=10)
-        
-    def open_login_window(self):
-        subprocess.Popen(["python3", "Login.py"], shell=False)
 
     def check_answer(self):
         username = self.username_entry.get()
@@ -146,8 +136,3 @@ class PasswordResetApp:
             self.set_new_password(username)
         else:
             self.show_message_box("Error", "Invalid Username or Security Question/Answer.")
-
-if __name__ == "__main__":
-    root = tk.Tk()
-    app = PasswordResetApp(root)
-    root.mainloop()
